getting started: remove everything related to updates
It's the primary source of confusion for new users, for example
when there are no updates, or updates require confirmations.

Instead depend on the installer being regularely updated and just
link to the update guide at the end.

Also try to nudge users into opening a mingw64 terminal, so they
are in the right environment right away. We should probably
do that automatically and also unify the naming of the environments,
now that there are so many of them.