getting-started: only suggest installing gcc for starters
The group install is confusing for beginners since it lists a lot
of packages and asks users to select which packages to install, which
they likely are not sure about.

Since we just want to show how to install things and give the user
some simple thing to achive just suggest installing gcc.

And since we have moved other groups like base/base-devel to meta packages
the knowledge of how groups work might not even be needed later on anyway.