Update to mkdocs-material v9
title handling seems to have changed:

* "title:" no longer gets used for h1
* having a html h1 on the home page no longer removes the original one

So convert all "title:" to real headings and add some inline css to
hide the orginal home page title.

The remaining config changes enable some features that were changed to opt-in.

The prev/next buttons at the bottom are also opt-in now, but I'm not missing them,
so leave them off for now.