Add a new page for the MSYS2 installer
This information up to now lived in the msys2-installer repo.
Move it here and expand it with some information to verify the
checksums and the signature.

With the signature/checksum stuff now being there we can just link it
from the main page. Most users wont know what signatures are, so let's not
confuse them with it.

The installer page now links to the hosted checksum file instead of embedding
it in the website, which in theory would reduce security, but if users want
to be sure they need to use gnupg anyway.

The signature part is a bit of a chicken egg thing since it assumes you have
gnupg available. Not sure how that could be improved.